Take k argument to get 'a's prefixed
'a's are treated as zeroes, so normally no distinction is drawn between
't' and 'at' and 'aaaaat' for the same reason that 3 and 03 and 000003
represent the same number.

Perhaps confusingly, the k value will be ignored if the resulting string
is already longer than k. It seems wrong to truncate it, but it also
seems wrong to have this behavior, which makes me think there's
something fundamentally conceputally wrong with specifying these
functions this way to begin with... but whatever, I also anticipate
finding out that there's a much neater way to implement this stuff in
Haskell as well.